+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changelog for NeoFS Node
 - SN could panic on forwarding GET/HEAD/RANGE request (#4120)
 - Storage node shuts down when SIGHUP configuration reload fails (#4113)
 - Storage node shuts down gracefully on invalid node attributes during SIGHUP (#4113)
+- SN exits on bootstrap if the FS chain mempool is full (#4128)
 
 ### Changed
 - SNs exchange TLS certificates on inter-node connections (#4097)